Update Add CodeFixes Analyser
Separate CodeFixes from SourceGenerators
Fix for #99
