+++ src/ReactiveUI.SourceGenerators.CodeFixers/AnalyzerReleases.Unshipped.md
@@ -1,0 +1,2 @@
+; Unshipped analyzer releases
+; https://github.com/dotnet/roslyn-analyzers/blob/main/src/Microsoft.CodeAnalysis.Analyzers/ReleaseTrackingAnalyzers.Help.md
