+++ src/ReactiveUI.SourceGenerators.CodeFixers/Diagnostics/CodeAnalyzers/AsyncVoidReturningReactiveCommandMethodAnalyzer.cs
@@ -9,7 +9,7 @@ using System.Linq;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Diagnostics;
 using ReactiveUI.SourceGenerators.Extensions;
-using static ReactiveUI.SourceGenerators.Diagnostics.DiagnosticDescriptors;
+using static ReactiveUI.SourceGenerators.CodeFixers.Diagnostics.DiagnosticDescriptors;
 
 namespace ReactiveUI.SourceGenerators;
