+++ src/ReactiveUI.SourceGenerators.CodeFixers/Diagnostics/SuppressionDescriptors.cs
