+++ src/ReactiveUI.SourceGenerators.CodeFixers/Diagnostics/Suppressions/OAPHMethodDoesNotNeedToBeStaticDiagnosticSuppressor.cs
