+++ src/ReactiveUI.SourceGenerators.CodeFixers/Diagnostics/Suppressions/ReactiveCommandMethodDoesNotNeedToBeStaticDiagnosticSuppressor.cs
